Begining of explicit list
I started implementing an explicit list based on the start code provider by dr. back.

The first thing that was changed was the block struct. A list_elem was added to hold the pointer for the prev and next elements. The min. block size was increased to 8 words. Since list_elem adds 2 words making the total block size 6 words. The size is pushed to 8 words for aligment.

The explicit follow s a very simple algorithm. Whenever a block is getting freed it's added to the list if it's getting allocatted it's removed from the free list. When coalescing there can be at most three blocks in question. The left most one if always kept in the explicit list with the reset being popped.

more updates needed. Performance tests didn't yeild that big of a difference.