removed list_elem from struct block
The block struct had a sturct list_elem paramter. The paramter held next and previous pointer to prev and next free blocks. However this was not needed as they could be stored in the payload paramter. Since free blocks don't have data in their payloads the payload can just hold the next and prev pointers.

I added the macro LIST_ELEM_TO_BLOCK that converts a list_elem to it's corresponding block. This is used when iterating through the list to get find a free block. This macro is a copy of the list_entry macro with a cast to cast payload to list_elem* as it's a char*