new block structure and optimized splitting
2 major changes
1 new block structure

instead of having the same block strucutre for free and used blocks, blocks were split into 2 types. The first type is block that holds a header and payload. This used blocked type will be used for dealing with used blocks. The second type is free_block that holds a header and a list_elem.

This had no eefect on utilization or throughput

2 optimized placing

placing an allocated portion of memory in a free block has 2 cases.
 case 1 that the allocated block fits the free block or when allocated would result in a free block smaller than the min block size. In this case that free block is fully allocated

case 2 the allocated block fits a portion of memory and the remaining free block is greater than the min block size. In this case there are two ways to split. Split and allocate to the right or split an allocate to the left.
	Split and allocate to the right
		in this case the fre_block is removed from the list (the allocated block would take place of it's list pointers)
		the block is marked with as used
		the remaining portion of the free block is reinserted into the list
	Split and allocate to the left:
		-free_blocked header updated with the smaller size
		-the remaining portion of the free block is allocated

in the second case we avoided removing and readding to the list. This would likely increase throughput

from testing this increased utilization and decreased throughput a little. (weird)