fixed some bugs in seg list and added a seg_list macro
First of all I added a SEG_LIST macro so we cna switch between implicit, explicit and seg list implmenetation quickly. The purpose of this is to be able to quickyl switch implementations for performance comparisions. In the previous commit the code for seg_list was randomly placed within EXPLICIT_LIST code which made no sense and ruined the structure of the code

Further more the seg list implemintation that was writen was half assed and not tested. where do I start.......

Intially in mm_init in the initilization of the lists the size that each list holds was not added correctly which resulted in all the blocks going to one list. The code was ran as if it was an explicit list implementation. (with proper testing that would have been obvious)

Secondly, the seg list implimentation doesn't handle coalescing. Hence blocks that are coalesed end up in the same list even if there belong to a different list when coalesced. (simple testing would have shown that utilization has signficantly decreased and raised the question of why that would be happening..)

lastly, the seg list implementation doesn't handle spliting. When a block splits it remains in the same list even if it's size now belongs to a different list. (simple testing would have shown that util
ization has signficantly decreased and raised the question of why that wo
uld be happening..)

Sloppy implementation if you ask me and apparently no testing was conducted.

Next step is to clean up this mess.