fixing the seg_list implemintation
I have fixed and updated the seg_list implementation. The main 2 additions were in the coalescing and placing.

Coalescing

During coalescing the block to the right is removed from it's list. Then we check if the size of the block after coleasing would fit in the left's block's list. If it would then it is kept in that list else it is put in a list it fits.

Splitting

When a block is split is size decreases.Hence it is removed from it's list and is placed in an approriorte sized list