Optimizted realloc case 2 for seg_list
My focus for this commit is optimization of a comibination of case 2 and case 3. This optimization was for a block being realloced with the next block being free irrespective of the state of the prev block.

If the next block to the one being reallocated is free and it's size is greater than the addition needed to the realloced block then that part of the next block is used. The size of the next block shrinks and it's then removed from it's list and added to one of approriate size.